+++ flutter_color_models/lib/src/models/cmyk_color.dart
@@ -113,44 +113,25 @@ class CmykColor extends cm.CmykColor
     return copyWith(alpha: (opacity * 255).round());
   }
 
-  // We must override this withValues from Color, has this from dart ui
-  /// Returns a new color that matches this color with the passed in components
-  /// changed.
-  ///
-  /// Changes to color components will be applied before applying changes to the
-  /// color space.
-  Color withValues(
-      {double? alpha,
-      double? red,
-      double? green,
-      double? blue,
-      ui.ColorSpace? colorSpace}) {
-    Color? updatedComponents;
-    if (alpha != null || red != null || green != null || blue != null) {
-      updatedComponents = Color.from(
-          alpha: alpha ?? a,
-          red: red ?? r,
-          green: green ?? g,
-          blue: blue ?? b,
-          colorSpace: this.colorSpace);
-    }
-    if (colorSpace != null && colorSpace != this.colorSpace) {
-      final UICloned_ColorTransform transform =
-          UICloned_getColorTransform(this.colorSpace, colorSpace);
-      return transform.transform(updatedComponents ?? this, colorSpace);
-    } else {
-      return updatedComponents ?? this;
-    }
-  }
-  
   @override
-  CmykColor withValuesList(List<num> values) {
+  CmykColor fromValues(List<num> values) {
     assert(values.length == 4 || values.length == 5);
     assert(values[0] >= 0 && values[0] <= 100);
     assert(values[1] >= 0 && values[1] <= 100);
     assert(values[2] >= 0 && values[2] <= 100);
     assert(values[3] >= 0 && values[3] <= 100);
     return CmykColor.fromList(values);
+  }
+
+  @override
+  CmykColor withValues(
+      {double? alpha,
+      double? red,
+      double? green,
+      double? blue,
+      ui.ColorSpace? colorSpace}) {
+    Color color = performWithValues(alpha, red, green, blue, colorSpace);
+    return CmykColor.fromColor(color);
   }
 
   @override
@@ -181,7 +162,7 @@ class CmykColor extends cm.CmykColor
   /// {@macro color_models.CmykColor.from}
   factory CmykColor.from(cm.ColorModel color) => color.toCmykColor().cast();
 
-  /// {@macro color_models.CmykColor.fromList}
+  /// {@macro color_models.CmykColor.fromValues}
   factory CmykColor.fromList(List<num> values) {
     assert(values.length == 4 || values.length == 5);
     assert(values[0] >= 0 && values[0] <= 100);
@@ -248,25 +229,5 @@ class CmykColor extends cm.CmykColor
 
   @override
   CmykColor convert(cm.ColorModel other) => other.toCmykColor().cast();
-
-  //OVERRIDEs for painting.dart Color
-  @override
-  double get a => (alpha / 255);
-
-  /// The red channel of this color.
-  @override
-  double get r =>(red / 255);
-
-  /// The green channel of this color.
-  @override
-  double get g => (green / 255);
-
-  /// The blue channel of this color.
-  @override
-  double get b => (blue / 255);
-
-  /// The color space of this color.
-  @override
-  final ui.ColorSpace colorSpace=ui.ColorSpace.sRGB;
-    
+  
 }