+++ flutter_color_models/lib/src/models/helpers/rgb_getters.dart
@@ -9,4 +9,12 @@ mixin RgbGetters on ColorModel {
 
   /// The blue value of this color.
   int get blue => toRgbColor().blue;
+
+  double get a => alpha / 255.0;
+
+  double get r => red / 255.0;
+
+  double get g => green / 255.0;
+
+  double get b => blue / 255.0;
 }